+++ lib/src/serializer/utils.dart
@@ -3,17 +3,21 @@ import 'repo.dart';
 typedef String KeyMaker<Key>(Key key);
 typedef dynamic ValueMaker<Value>(Value value);
 
-Map<String, dynamic> mapMaker<Key, Value>(
-    Map map, KeyMaker<Key> keyMaker, ValueMaker<Value> valueMaker) {
+Map<String, dynamic> nullableMapMaker<Value>(
+    Map<String, dynamic> map, ValueMaker<Value> valueMaker) {
   Map<String, dynamic> ret;
   if (map != null) {
     ret = <String, dynamic>{};
-    map.forEach((key, value) {
-      ret[keyMaker(key)] = valueMaker(value);
-    });
+    for (final key in map.keys) {
+      ret[key] = valueMaker(map[key]);
+    }
   }
   return ret;
 }
+
+Map<String, dynamic> nonNullableMapMaker<Value>(Map<String, dynamic> map,
+        ValueMaker<Value> valueMaker, Map<String, dynamic> defaultValues) =>
+    nullableMapMaker<Value>(map, valueMaker) ?? defaultValues;
 
 void setNullableValue(Map<String, dynamic> map, String key, value) {
   map[key] = value;
@@ -21,17 +25,16 @@ void setNullableValue(Map<String, dynamic> map, String key, value) {
 
 void setNonNullableValue(Map<String, dynamic> map, String key, value) {
   if (value != null) {
-    map[key] = value;
+    setNullableValue(map, key, value);
   }
 }
 
-List safeIterableMapper<T>(Iterable<T> values, callback(T value)) =>
-    values?.map((T value) {
-      if (value == null) {
-        return null;
-      }
-      return callback(value);
-    })?.toList();
+List nullableIterableMapper<T>(Iterable<T> values, callback(T value)) =>
+    values?.map(callback)?.toList();
+
+List nonNullableIterableMapper<T>(
+        Iterable<T> values, callback(T value), List<T> defaultValues) =>
+    nullableIterableMapper<T>(values, callback) ?? defaultValues;
 
 void setTypeKeyValue(String typeKey, String modelString, bool withType,
     Map<String, dynamic> map) {